windows stuff

+++ scilab/modules/colpack/sci_gateway/cpp/sci_spCompHessian.cpp
@@ -73,7 +73,7 @@ types::Function::ReturnValue sci_spCompHessian(types::typed_list &in, types::opt
 
     vector<int> viColors;
     spchEngine->getColumnColors(viColors);
-    types::Double *pDblCol = new types::Double(viColors.size(),1);
+    types::Double *pDblCol = new types::Double((int)viColors.size(),1);
     for (int i=0; i<pDblCol->getSize(); i++)
     {
         pDblCol->set(i,1+viColors[i]);
